agregando seeders

+++ database/migrations/2022_05_02_121112_create_anios_table.php
@@ -15,7 +15,7 @@ class CreateAniosTable extends Migration
     {
         Schema::create('anios', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('anio')->unique();
+            $table->integer('anio')->unique();
         });
     }
